src/app.py
import pandas as pd
import numpy as np
import hydra
import os
import json
import pickle
import time

from flask import request, jsonify

#5.6.2021
#1. using the selected model to predict the new data
#2. build the api (get, post) by using flask
#3. exception handling